Moved all mocks to acceptor_test.
