--- network/acceptor_mock.go
@@ -1,39 +1,0 @@
-// Copyright (c) 2017 The Alvalor Authors
-//
-// This file is part of Alvalor.
-//
-// Alvalor is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Alvalor is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with Alvalor.  If not, see <http://www.gnu.org/licenses/>.
-
-package network
-
-import (
-	"github.com/stretchr/testify/mock"
-	"net"
-)
-
-type acceptorMock struct {
-	mock.Mock
-}
-
-func (acceptor *acceptorMock) ClaimSlot() error {
-	args := acceptor.Called()
-	return args.Error(0)
-}
-func (acceptor *acceptorMock) ReleaseSlot() {
-	acceptor.Called()
-}
-func (acceptor *acceptorMock) AddPeer(conn net.Conn, nonce []byte) error {
-	args := acceptor.Called(conn, nonce)
-	return args.Error(0)
-}
